requirements.txt
requests
xmltodict
pyyaml